1. CY8PROTO_062_4343W - Device user led is 13.7 2. Previously cm0 - core was controlling the P13.7, now using psoc62_sleep function to put cm0 in sleep mode 3. Basic Led toggle is working - delay is controlled by global variable

+++ src/uart.c
@@ -2,7 +2,6 @@
 
 const char *string = "HelloWorld : ";
 
-
 /* Allocate context for UART operation */
 cy_stc_scb_uart_context_t uartContext;
 /* Populate configuration structure */
@@ -34,8 +33,7 @@ const cy_stc_scb_uart_config_t uartConfig = {
     .txFifoIntEnableMask = 0UL,
 };
 
-uint32_t count = 0;
-
+uint8_t count = 0;
 
 void uart_init()
 {
@@ -62,13 +60,12 @@ void uart_init()
     Cy_SCB_UART_Init(UART, &uartConfig, &uartContext);
 
     Cy_SCB_UART_Enable(UART);
-
 }
 
 void uart_transmit()
 {
 
-	Cy_SCB_UART_PutString(UART, string);
-        Cy_SCB_UART_Put(UART, count++);
-        Cy_SCB_UART_PutString(UART, "\n");
+    Cy_SCB_UART_PutString(UART, string);
+    Cy_SCB_UART_Put(UART, count++);
+    Cy_SCB_UART_PutString(UART, "\n");
 }
